Some double->float conversion

+++ trunk/sfox/src/starengine/engine/camera.h
@@ -13,7 +13,7 @@ typedef struct camera {
 
   frustum ftm;                  /* Initialised in camera_to_opengl */
   viewport vp;
-  double fov, znear, zfar;
+  float fov, znear, zfar;
   matrix4 projection_matrix;
 
   unsigned int need_update;
@@ -24,21 +24,21 @@ typedef struct camera {
 
 #define CAM_UNLIMITED -1
 
-extern camera camera_create(double fov, double zfar, double znear, vector3 *pos, vector3 *up, viewport vp);
+extern camera camera_create(float fov, float zfar, float znear, vector3 *pos, vector3 *up, viewport vp);
 /* Destroy the viewport */
 extern void camera_destroy(camera cam);
 extern void camera_to_opengl(camera cam);
 
-extern void camera_set_pos(camera cam, double x, double y, double z);
+extern void camera_set_pos(camera cam, float x, float y, float z);
 
 /*Function for fps-like camera */
 
 /* Rotate the camera of xrel and yrel radians around x and y axes*/
-extern void camera_mouse_move(camera cam, double xrel, double yrel);
+extern void camera_mouse_move(camera cam, float xrel, float yrel);
 /* Move the camera in the view direction at speed speed*/
-extern void camera_move_along_view(camera cam, double speed);
+extern void camera_move_along_view(camera cam, float speed);
 /*Strafe*/
-extern void camera_side_move(camera cam, double speed);
+extern void camera_side_move(camera cam, float speed);
 
 extern void camera_set_max_ypr(camera cam, float yaw, float pitch, float roll);
